remi/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
twilio==9.3.0
python-multipart==0.0.9
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
